Friday Cancelation
Friday cancelation is meant to be scheduled to run on a Friday at 12:00am.
It checks who is signed up for a session and if the numbers are too small the session is "canceled" (nothing removed as of yet) and those who need to be are notified.

Fixed an Issue with course_signup where it would redirect to an older version.
added blocking in session conformation to prevent signing up for past sessions or sessions monts in advance.