added mentor page and session conformation
Session conformation is now located on the dashboard.

the link in course singup is fixed, andthe sql close is now located at the bottom of the page.,

Friday cancelation now takes the super mod parameter. which will allow for more flexability in adding high level mods.

Moderato allows a moderator to sign up to moderate a class. and shows the name of the user who is moderating the class currently.

session_conformation( soon deprecated) now actually changes the date instead of both variables refferencing the same value.