Completed register_parent functionality
Completed register_parent activity in app. Activity sends form info to PHP script and receives a JSON object with the result; if user is registered without errors, the app shows a toast alerting the user and returns to the parent activity; otherwise, it shows a toast with the first error found.