Addition of view sections page
View Sections allows a user to view eachclass in the school, this offers a drop down showing each section of the class offered, This in turn offers a drop down showing information on that specific section.

This update also retools the all_classes api call in order to more efficently store the data.